add pinia
