+++ src/main.ts
@@ -1,5 +1,7 @@
 import 'virtual:windi-devtools'
 import 'virtual:windi.css'
+
+import { createPinia } from 'pinia'
 import { createApp } from 'vue'
 
 import App from './App.vue'
@@ -7,4 +9,4 @@ import router from './router'
 
 const app = createApp(App)
 
-app.use(router).mount('#app')
+app.use(router).use(createPinia()).mount('#app')
