+++ src/pages/Requests.vue
@@ -1,50 +1,11 @@
 <script lang="ts" setup>
+import { storeToRefs } from 'pinia'
 import { ref } from 'vue'
 import FilteringMenu from '../components/FilteringMenu.vue'
 import Request from '../components/Request.vue'
-const requests = [
-  {
-    id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
-    status: 'submitted',
-    created_at: '2022-01-25T13:29:19.918Z',
-    updated_at: '2022-01-25T13:29:19.918Z',
-    created_by: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
-    amount: 1200,
-    title: 'SysAd講習会の開催費用',
-    content: 'サーバー代 1200円',
-    tags: [
-      {
-        id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
-        name: '2020講習会',
-        description: '2020年度講習会',
-        created_at: '2022-01-25T13:29:19.918Z',
-        updated_at: '2022-01-25T13:29:19.918Z'
-      },
-      {
-        id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
-        name: '2020講習会',
-        description: '2020年度講習会',
-        created_at: '2022-01-25T13:29:19.918Z',
-        updated_at: '2022-01-25T13:29:19.918Z'
-      },
-      {
-        id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
-        name: '2020講習会',
-        description: '2020年度講習会',
-        created_at: '2022-01-25T13:29:19.918Z',
-        updated_at: '2022-01-25T13:29:19.918Z'
-      }
-    ],
-    group: {
-      id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
-      name: 'SysAd',
-      description: 'SysAd班',
-      budget: 250000,
-      created_at: '2022-01-25T13:29:19.918Z',
-      updated_at: '2022-01-25T13:29:19.918Z'
-    }
-  }
-]
+import { useRequestStore } from '../stores/request'
+const requestStore = useRequestStore()
+const { requests } = storeToRefs(requestStore)
 const isOpen = ref(false)
 function open() {
   isOpen.value = !isOpen.value
