+++ src/stores/request.ts
@@ -41,7 +41,60 @@ type Params = {
 
 export const useRequestStore = defineStore('request', {
   state: () => ({
-    requests: new Array<Request>()
+    requests: [
+      {
+        id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
+        status: 'submitted',
+        created_at: '2022-01-25T13:29:19.918Z',
+        updated_at: '2022-01-25T13:29:19.918Z',
+        created_by: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
+        amount: 1200,
+        title: 'SysAd講習会の開催費用',
+        content: 'サーバー代 1200円',
+        tags: [
+          {
+            id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
+            name: '2020講習会',
+            description: '2020年度講習会',
+            created_at: '2022-01-25T13:29:19.918Z',
+            updated_at: '2022-01-25T13:29:19.918Z'
+          },
+          {
+            id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
+            name: '2020講習会',
+            description: '2020年度講習会',
+            created_at: '2022-01-25T13:29:19.918Z',
+            updated_at: '2022-01-25T13:29:19.918Z'
+          },
+          {
+            id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
+            name: '2020講習会',
+            description: '2020年度講習会',
+            created_at: '2022-01-25T13:29:19.918Z',
+            updated_at: '2022-01-25T13:29:19.918Z'
+          }
+        ],
+        group: {
+          id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
+          name: 'SysAd',
+          description: 'SysAd班',
+          budget: 250000,
+          created_at: '2022-01-25T13:29:19.918Z',
+          updated_at: '2022-01-25T13:29:19.918Z'
+        }
+      }
+    ], //new Array<Request>()
+    params: {
+      sort: 'created_at',
+      current_state: '',
+      year: '',
+      target: '',
+      type: '',
+      since: '',
+      until: '',
+      tag: '',
+      group: ''
+    } as Params
   }),
   getters: {
     dateFormatter() {
@@ -56,9 +109,24 @@ export const useRequestStore = defineStore('request', {
     }
   },
   actions: {
-    async getRequests(params: Params) {
-      const response: Request[] = await axios.get('/api/requests', { params })
+    async getRequests() {
+      const response: Request[] = await axios.get('/api/requests', {
+        params: this.params
+      })
       this.requests = response
+    },
+    resetParams() {
+      this.params = {
+        sort: 'created_at',
+        current_state: '',
+        year: '',
+        target: '',
+        type: '',
+        since: '',
+        until: '',
+        tag: '',
+        group: ''
+      }
     }
   }
 })
